+++ lib/instream/connection/config.ex
@@ -58,6 +58,10 @@ defmodule Instream.Connection.Config do
     end
   end
 
+  defp maybe_fetch_system({ :system, var, default }) do
+    System.get_env(var) || default
+  end
+
   defp maybe_fetch_system({ :system, var }), do: System.get_env(var)
   defp maybe_fetch_system(config),           do: config